software/py/requirements.txt
aenum>=2.1.4
astroid==2.4.2
bitarray>=1.1.0
colorama>=0.4.1
epc>=0.0.5
fire>=0.1.3
flask>=1.1.1
imageio>=2.4.1
ipython-genutils>=0.2.0
ipython>=7.1.1
jedi>=0.13.1
jinja2>=2.10.3
jsonpickle>=1.0
matplotlib>=3.0.2
numba>=0.51.2
numpy>=1.16.2
pandas>=0.24.2
progressbar2>=3.53.1
pylint===2.6.0
pyserial>=3.4
python-can>=2.2.0
pyusb>=1.0.2
pyyaml>=5.1.2
requests>=2.23.0
scipy>=1.4.0
seaborn>=0.11.2
timeout-decorator>=0.4.0
tqdm==4.65.0
virtualenv>=16.6.2
watchdog>=0.9.0
typing_extensions>=3.10.0.0
